--- client/src/HorseInput.ts
@@ -1,12 +1,0 @@
-// HorseInput.ts
-interface HorseInput {
-  name: string;
-  age: number;
-  breed: string;
-  racingStatus: string;
-  tokenURI: string;
-  imageURL: string;
-  saleType: string;
-  price: number;
-  deadline: Date | null;
-}
