+++ client/src/lib/controllers/HorseMarketController.ts
@@ -1,17 +1,15 @@
 import EthersProvider from "/workspace/Block-Gallop-Stables/client/src/lib/providers/ethersProvider";
-
+import MetamaskController from "/workspace/Block-Gallop-Stables/client/src/lib/controllers/MetamaskController";
+import { RoleManagerController } from "/workspace/Block-Gallop-Stables/client/src/lib/controllers/RoleManagerController";
 
 export class HorseMarketController {
-	account(account: any) {
-		throw new Error('Method not implemented.');
-	}
+  roleManagerController: RoleManagerController;
   ethersProvider: EthersProvider;
   horseMarketContract: any;
-static listHorseForSale: any;
-static init: any;
 
   constructor() {
     this.ethersProvider = new EthersProvider();
+    this.roleManagerController = new RoleManagerController();
   }
 
   async init() {
@@ -20,16 +18,24 @@ static init: any;
   }
 
   async listHorseForSale(tokenId: number, saleType: number, price: number, deadline: number) {
-    console.log("Calling listHorseForSale");
-    console.log(`tokenId: ${tokenId}, saleType: ${saleType}, price: ${price}, deadline: ${deadline}`);
-    try {
-      const result = await this.horseMarketContract.listHorseForSale(tokenId, saleType, price, deadline);
-      console.log("List horse for sale result: ", result);
-      return result;
-    } catch (error) {
-      console.error("Error listing horse for sale: ", error);
-      throw error;
-    }
+      console.log("Calling listHorseForSale");
+      console.log(`tokenId: ${tokenId}, saleType: ${saleType}, price: ${price}, deadline: ${deadline}`);
+      try {
+          await MetamaskController.init();
+          const metamaskControllerStore = MetamaskController.store;
+          if (!metamaskControllerStore.isConneted) {
+              console.error('Metamask is not connected to the right network');
+              return;
+          }
+          const account = metamaskControllerStore.activeAccount;
+          const result = await this.horseMarketContract.listHorseForSale(tokenId, saleType, price, deadline, { from: account });
+          console.log("List horse for sale result: ", result);
+          await this.roleManagerController.grantRoleToSeller(account);
+          return result;
+      } catch (error) {
+          console.error("Error listing horse for sale: ", error);
+          throw error;
+      }
   }
 }
